Working mobile-first, responsive website
A new structure was adopted for the HTML, where the header is followed
by a `main` section. This is further broken up into a series of
`section` blocks, each describing a component of the site. Though the
HTML is likely a final draft, the CSS still needs quite a bit of work
to make the site look and feel modern.