Remove old files, update log for 6/29/18
Since `index.html` and `style.css` aren't being used anymore, I took
them out of the repo. The `README` was also updated to reflect this.
The only thing that needs to change is the Codepen, however, I will
just wait until the project is done and update it then. Unless I
choose to move development back to there, in which case the files
here will contine to be snapshots as originally intended.