Update Pricing Cards
Designing for a 320px mobile viewport is hard because everything has
to be stacked. I chose a 'lightweight' style for this set of cards
because the previous set was completely filled in. One issue is this
is the first spot I use black text; not sure if that can be changed.