Update Contact Form
In keeping with the theme of getting lighter as you scroll down the
page, I made the contact form even lighter than the previous cards.
The sumbit button is supposed to be eye catching with the dark
background though. Also, I tried to make it inviting by saying
"Let's chat!" instead of something generic like "Submit".